Revert "merge upstream"

+++ Ryujinx.HLE/HOS/Services/Hid/HidDevices/Types/KeyboardInput.cs
@@ -3,6 +3,6 @@ namespace Ryujinx.HLE.HOS.Services.Hid
     public struct KeyboardInput
     {
         public int Modifier;
-        public ulong[] Keys;
+        public int[] Keys;
     }
 }
